requirements.txt
google-genai
tavily-python
discord.py
python-dotenv
requests
